// angular-frontend/src/app/models/Worker.ts
export class Worker{
    
    workerId: number;
    lastName: string;
    firstName: string;
    username: string;
    passCode: string;
    title: string;
    reportsTo: number;
    hireDate: string;
    streetAddress: string;
    city: string;
    state: string;
    zip: string;
    phone: string
    email: string;

    constructor() {}

}